+++ src/react/src/components/FacilityDetailsGeneralFields.jsx
@@ -9,11 +9,7 @@ import FacilityDetailsItem from './FacilityDetailsItem';
 import FacilityDetailsClaimedInfo from './FacilityDetailsClaimedInfo/FacilityDetailsClaimedInfo';
 import ShowOnly from './ShowOnly';
 import FeatureFlag from './FeatureFlag';
-import {
-    formatAttribution,
-    formatExtendedField,
-    formatPartnerFieldValue,
-} from '../util/util';
+import { formatAttribution, formatExtendedField } from '../util/util';
 
 import {
     EXTENDED_FIELD_TYPES,
@@ -231,37 +227,6 @@ const FacilityDetailsGeneralFields = ({
         );
     };
 
-    const renderPartnerField = ({
-        label,
-        fieldName,
-        formatValue,
-        partnerConfigFields,
-    }) => {
-        const values = get(data, `properties.partner_fields.${fieldName}`, []);
-
-        if (!values.length || !values[0]) return null;
-
-        const formatField = item =>
-            formatExtendedField({
-                ...item,
-                formatValue,
-            });
-
-        const topValue = formatField(values[0]);
-
-        return (
-            <Grid item xs={12} md={6} key={`partner-${label}`}>
-                <FacilityDetailsItem
-                    {...topValue}
-                    label={topValue.label ? topValue.label : label}
-                    additionalContent={values.slice(1).map(formatField)}
-                    embed={embed}
-                    partnerConfigFields={partnerConfigFields}
-                />
-            </Grid>
-        );
-    };
-
     const contributorFields = filter(
         get(data, 'properties.contributor_fields', null),
         field => field.value !== null,
@@ -297,47 +262,14 @@ const FacilityDetailsGeneralFields = ({
             field => !ADDITIONAL_IDENTIFIERS.includes(field.fieldName),
         );
 
-        const partnerFieldNames = Object.keys(
-            get(data, 'properties.partner_fields', {}),
-        );
-
-        const partnerFields = partnerFieldNames.map(fieldName => {
-            /* 
-            We have to rely on the first element of the partner-field list
-            because the backend isn’t configured to store partner-specific
-            settings in a separate metadata object.
-            */
-            const firstEntry =
-                get(data, `properties.partner_fields.${fieldName}[0]`) || {};
-            const {
-                base_url: baseUrl, // eslint-disable-line camelcase
-                display_text: displayText, // eslint-disable-line camelcase
-            } = firstEntry;
-
-            const partnerConfigFields = { baseUrl, displayText };
-
-            return {
-                fieldName,
-                label: fieldName
-                    .replace(/_/g, ' ')
-                    .replace(/\b\w/g, l => l.toUpperCase()),
-                formatValue: formatPartnerFieldValue,
-                partnerConfigFields,
-            };
-        });
-
         return (
             <FeatureFlag
                 flag={SHOW_ADDITIONAL_IDENTIFIERS}
-                alternative={[
-                    ...extendedFieldsWithoutAdditionalIdentifiers.map(
-                        renderExtendedField,
-                    ),
-                    ...partnerFields.map(renderPartnerField),
-                ]}
+                alternative={extendedFieldsWithoutAdditionalIdentifiers.map(
+                    renderExtendedField,
+                )}
             >
                 {EXTENDED_FIELD_TYPES.map(renderExtendedField)}
-                {partnerFields.map(renderPartnerField)}
             </FeatureFlag>
         );
     };