+++ src/react/src/util/constants.jsx
@@ -378,6 +378,7 @@ export const productionLocationInfoRouteCreate =
     '/contribute/single-location/info/:moderationID?';
 export const productionLocationInfoRouteUpdate =
     '/contribute/single-location/:osID/info/:moderationID?';
+export const pilotIntegrationsDocsRoute = `${InfoLink}/data-integrations`;
 
 export const contributeFieldsEnum = Object.freeze({
     name: 'name',